theming: Frontpage layout
* NEW HTML + CSS application layout.
* NEW Add Bootstrap 4 library.
* NEW Add Google Roboto font.
