+++ sonar/modules/documents/views.py
@@ -10,13 +10,14 @@
 
 from __future__ import absolute_import, print_function
 
-from flask import Blueprint
+from flask import Blueprint, g, render_template
 
 blueprint = Blueprint(
     'documents',
     __name__,
     template_folder='templates',
-    static_folder='static'
+    static_folder='static',
+    url_prefix='/organization/<ir>'
 )
 """Blueprint used for loading templates and static assets
 
@@ -24,3 +25,33 @@ The sole purpose of this blueprint is to ensure that Invenio can find the
 templates and static files located in the folders of the same names next to
 this file.
 """
+
+
+@blueprint.url_defaults
+def add_ir(endpoint, values):
+    """Add default ir parameter."""
+    values.setdefault('ir', 'sonar')
+
+
+@blueprint.url_value_preprocessor
+def pull_ir(endpoint, values):
+    """Add ir parameter to global variables."""
+    g.ir = values.pop('ir')
+
+
+@blueprint.route('/')
+def index():
+    """IR (and SONAR) home view."""
+    return render_template('sonar/frontpage.html')
+
+
+@blueprint.route('/search')
+def search():
+    """IR search results."""
+    return render_template('sonar/search.html')
+
+
+@blueprint.route('/detail/<pid>')
+def detail():
+    """Search details."""
+    return render_template('sonar/search.html')
