+++ sonar/theme/assets/scss/common/_variables.scss
@@ -4,3 +4,5 @@ $primary: #205078 !default;
 $secondary: rgb(246, 130, 17) !default;
 
 $bg-ir: $primary !default;
+
+$font-family-base: 'Roboto Condensed', sans-serif;
