+++ sonar/theme/assets/scss/sonar/theme.scss
@@ -12,6 +12,4 @@
 // @import "../invenio_theme/styles";
 
 @import "variables";
-@import "~bootstrap/scss/bootstrap";
-@import '~font-awesome/scss/font-awesome.scss';
 @import "../common/theme";
