+++ sonar/theme/assets/scss/usi/theme.scss
@@ -12,8 +12,6 @@
 // @import "../invenio_theme/styles";
 
 @import "variables";
-@import "~bootstrap/scss/bootstrap";
-@import '~font-awesome/scss/font-awesome.scss';
 
 .text-ir {
   color: $primary !important;
