+++ sonar/theme/templates/sonar/accounts/forgot_password.html
@@ -5,9 +5,9 @@
 
 {% block panel %}
   {%- with form = forgot_password_form %}
-    <div class="col-12 col-sm-6">
+    <div class="col-12 col-lg-6">
       <div class="card card-light text-dark">
-        <div class="card-body p-5">
+        <div class="card-body p-3 p-lg-5">
           <h4 class="text-center mb-4">{{_('Reset Password')}}</h3>
           {%- if messages %}
             {%- for category, message in messages %}
