+++ sonar/theme/templates/sonar/accounts/login.html
@@ -3,9 +3,9 @@
 {% from "invenio_accounts/_macros.html" import render_field, form_errors %}
 
 {% block panel %}
-  <div class="col-12 col-sm-6">
+  <div class="col-12 col-lg-6">
     <div class="card card-light text-dark">
-      <div class="card-body p-5">
+      <div class="card-body p-3 p-lg-5">
         <h4 class="text-center mb-4">{{_('Log in to account') }}</h3>
         <form action="{{ url_for_security('login') }}" method="POST" name="login_user_form">
           {%- with form = login_user_form %}
