+++ sonar/theme/templates/sonar/accounts/signup.html
@@ -4,9 +4,9 @@
 
 {% block panel %}
 
-<div class="col-12 col-sm-6">
+<div class="col-12 col-lg-6">
   <div class="card card-light text-dark">
-    <div class="card-body p-5">
+    <div class="card-body p-3 p-lg-5">
       <h4 class="text-center mb-4">{% trans sitename=config.ACCOUNTS_SITENAME %}Sign up for a {{ sitename }} account!{% endtrans %}</h3>
         {%- with form = register_user_form %}
           <form action="{{ url_for_security('register') }}" method="POST" name="register_user_form">
