+++ sonar/theme/templates/sonar/contact.html
@@ -1,0 +1,5 @@
+{% extends 'sonar/page.html' %}
+
+{%- block page_body %}
+    <h1>Contact</h1>
+{% endblock %}
