+++ sonar/theme/templates/sonar/footer.html
@@ -10,6 +10,7 @@
 <footer id="footer" class="bg-light mt-5">
   <div class="container">
     <div class="row justify-content-center align-items-center">
+      {#
       <div class="col-6 col-sm-6 py-2 py-sm-5">
         {%- if config.I18N_LANGUAGES %}
           <form id="language-code-form" class="form-inline lang-select"
@@ -24,8 +25,22 @@
           </form>
         {%- endif %}
       </div>
+      #}
+      <div class="col-12 col-sm-6 py-2 py-sm-5 text-center text-sm-left">
+        <div class="d-flex flex-row">
+          <div>
+              <a href="#">{{_('Help')}}</a>
+          </div>
+          <div class="ml-3">
+              <a href="#">{{_('About')}}</a>
+          </div>
+          <div class="ml-3">
+              <a href="#">{{_('Contact')}}</a>
+          </div>
+        </div>
+      </div>
       <div class="col-12 col-sm-6 py-2 py-sm-5 text-center text-sm-right">
-        {% trans link="https://rero.ch" %}Powered by <a href="{{link}}">RERO</a>{% endtrans %}
+        {% trans link="https://www.rero.ch" %}Powered by <a href="{{link}}" target="_blank">RERO</a>{% endtrans %}
       </div>
     </div>
   </div>
