--- sonar/theme/templates/sonar/header.html
@@ -1,71 +1,0 @@
-{# -*- coding: utf-8 -*-
-
-  This file is part of Invenio.
-  Copyright (C) 2015-2018 CERN.
-
-  Invenio is free software; you can redistribute it and/or modify it
-  under the terms of the MIT License; see LICENSE file for more details.
-#}
-
-<header>
-  <nav class="navbar navbar-expand-lg navbar-dark bg-dark">
-    <div class="container">
-      <button class="navbar-toggler ml-auto" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
-        <i class="fa fa-bars fa-lg p-2"></i>
-      </button>
-      <div class="collapse navbar-collapse" id="navbarSupportedContent">
-        {% if g.ir|default('sonar') != 'sonar' %}
-          <ul class="navbar-nav mr-auto">
-            <li class="nav-item">
-              <a class="nav-link" href="{{url_for('invenio_theme_frontpage.index')}}">
-                {{_('Back to SONAR')}}
-              </a>
-            </li>
-          </ul>
-        {% endif %}
-        {%- include 'sonar/header_login.html' %}
-      </div>
-    </div>
-  </nav>
-  {%- block navbar %}
-    <div class="bg-ir p-4">
-      <div class="container">
-        <div class="row justify-content-center">
-          <div class="col-8 col-xs-6 col-lg-4 py-3 py-sm-5">
-              <a href="{{ url_for('sonar.index', ir=g.ir|default('sonar')) }}">
-                  {%- if config.THEME_LOGO %}
-                    <img src="{{ url_for('static', filename='images/' ~ g.ir|default('sonar') ~ '-logo.svg')}}" alt="{{_(config.THEME_SITENAME)}}" class="logo"/>
-                  {%- else %}
-                    {{_(config.THEME_SITENAME)}}
-                  {%- endif %}
-                  </a>
-          </div>
-        </div>
-        {%- if config.THEME_SEARCHBAR %}
-        <div class="row justify-content-center">
-          <div class="col-sm-12 col-lg-8 text-right my-4">
-            <form class="justify-content-end" action="{{ url_for('sonar.search', ir=g.ir|default('sonar')) }}" role="search">
-              <div class="row">
-                <div class="col-12 col-sm-11">
-                    <input class="form-control form-control-lg mr-2" type="search" placeholder="{{_('Search publications, authors, projects, ...')}}" aria-label="Search" name="q">
-                    <div class="text-right mt-2 mr-2">
-                      <a href="#" class="text-light"><u>> {{_('Advanced search')}}</u></a>
-                    </div>
-                </div>
-                <div class="col-2 col-sm-1 d-none d-sm-block">
-                  <button class="btn btn-link btn-lg" type="submit">
-                    <i class="fa fa-search fa-lg text-ir"></i>
-                  </button>
-                </div>
-              </div>
-            </form>
-          </div>
-        </div>
-        {% endif %}
-      </div>
-    </div>
-  {%- endblock navbar %}
-    <div class="bg-secondary text-light text-center py-3">
-      <h5 class="m-0">Software under development !</h6>
-    </div>
-</header>
