+++ sonar/theme/templates/sonar/page.html
@@ -6,6 +6,9 @@
   Invenio is free software; you can redistribute it and/or modify it
   under the terms of the MIT License; see LICENSE file for more details.
 #}
+
+{% set page = page|default('int') %}
+
 <!DOCTYPE html>
 <html {% if html_css_classes %} class="{{ html_css_classes|join(' ') }}"{% endif %} lang="{{ current_i18n.locale.language|safe }}" dir="{{ current_i18n.locale.text_direction }}">
   <head>
@@ -54,8 +57,6 @@
         {%- endblock head_apple_icons %}
   
       {%- endblock head_links %}
-      
-      {%- block header %}{% endblock header %}
     
       {%- block css %}
         {{ webpack[g.ir|default('sonar') ~ '-theme.css'] }}
@@ -78,7 +79,10 @@
       {%- endblock browserupgrade %}
         
       {%- block body_inner %}
-        {% include 'sonar/header.html' %}
+          <header>
+              {% include 'sonar/partial/navbar.html' %}            
+              {%- block header %}{%- endblock header %}
+          </header>
           
         <div class="container my-5">
           {%- block breadcrumbs %}
