+++ sonar/theme/templates/sonar/partial/dropdown_user.html
@@ -1,0 +1,9 @@
+<div class="dropdown-menu dropdown-menu-right" aria-labelledby="{{dropdownId}}">
+    <h6 class="dropdown-header">{{current_user.email}}</h6>
+{%- if config.USERPROFILES %}
+    <a class="dropdown-item" href="{{ url_for('invenio_userprofiles.profile')}}">
+        {{_('Profile')}}
+    </a>
+{% endif %}
+    <a class="dropdown-item" href="{{url_for_security('logout')}}">{{_('Logout')}}</a>
+</div>
