+++ sonar/theme/templates/sonar/partial/navbar.html
@@ -1,0 +1,82 @@
+<nav class="navbar navbar-expand-lg navbar-dark bg-{{ 'dark' if page == 'home' else 'primary' }}">
+    <div class="container">
+        {%- if not current_user.is_authenticated %}
+            <a href="{{url_for_security('login', next=request.path)}}" class="btn btn-light ml-auto mr-2 d-block d-lg-none">
+                <i class="fa fa-lock"></i>
+            </a>
+        {% else %}
+            <div class="dropdown ml-auto mr-2 d-block d-lg-none">
+            {% with dropdownId='userButtonDropdown' %}
+                <button class="btn btn-light" type="button" id="{{dropdownId}}" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+                    <i class="fa fa-user"></i>
+                </button>
+                {% include 'sonar/partial/dropdown_user.html' %}
+            {% endwith %}
+            </div>
+        {% endif%}
+        <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
+            <i class="fa fa-bars p-1"></i>
+        </button>
+        <div class="collapse navbar-collapse" id="navbarSupportedContent">
+            {% if page != 'home' %}
+            <a class="navbar-brand" href="{{ url_for('documents.index', ir=g.ir|default('sonar')) }}">
+                <img src="{{ url_for('static', filename='images/' ~ g.ir|default('sonar') ~ '-logo.svg')}}" height="50" class="d-inline-block align-top mr-3 my-4" alt="">
+            </a>
+            <form action="{{ url_for('documents.search', ir=g.ir|default('sonar')) }}" class="form-inline my-2 my-lg-0 ml-lg-5">
+                <input name="q" class="form-control mr-sm-2" type="search" placeholder="{{_('Search')}}" aria-label="{{_('Search')}}">
+                <button class="btn btn-outline-light my-2 my-sm-0" type="submit">
+                    <i class="fa fa-search"></i>
+                </button>
+            </form>
+            {% endif %}
+            {% if g.ir|default('sonar') != 'sonar' %}
+                <ul class="navbar-nav mr-auto ml-4">
+                    <li class="nav-item">
+                        <a class="nav-link" href="{{url_for('invenio_theme_frontpage.index')}}">
+                            {{_('Back to SONAR')}}
+                        </a>
+                    </li>
+                </ul>
+            {% endif %}
+            <ul class="navbar-nav ml-auto">
+                {%- if not current_user.is_authenticated %}
+                    <li class="nav-item px-3 d-none d-lg-block">
+                        <a class="nav-link" href="{{url_for_security('login', next=request.path)}}" title="{{_('Log in')}}">
+                            <i class="fa fa-lock fa-lg"></i>
+                        </a>
+                    </li>
+                    {#
+                    {% if security.registerable %}
+                    <li class="nav-item">
+                        <a class="nav-link" href="{{url_for_security('register')}}">
+                        <i class="fa fa-sign-in fa-lg"></i> {{_('Sign up')}}
+                        </a>
+                    </li>
+                    {% endif %}
+                    #}
+                {% else %}
+                    <li class="nav-item dropdown d-none d-lg-block px-3">
+                    {% with dropdownId='userLinkDropdown' %}
+                        <a class="nav-link" href="#" id="{{dropdownId}}" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+                            <i class="fa fa-user"></i>
+                        </a>
+                        {% include 'sonar/partial/dropdown_user.html' %}
+                    {% endwith %}
+                    </li>
+                {% endif %}
+                {%- if config.I18N_LANGUAGES %}
+                    <li class="nav-item dropdown pl-3">
+                        <a class="nav-link" href="#" id="languageDropdown" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+                            {{ current_i18n.language|upper }}
+                        </a>
+                        <div class="dropdown-menu dropdown-menu-right" aria-labelledby="languageDropdown">
+                        {% for l in current_i18n.get_locales() %}
+                            <a class="dropdown-item" href="#">{{ l.get_display_name()|capitalize }}</a>
+                        {% endfor %}
+                        </div>
+                    </li>
+                {% endif %}
+            </ul>
+        </div>
+    </div>
+</nav>
